Update Chat Utils

+++ Dtos/ChatGroupDto/RecentActiveDto.cs
@@ -8,7 +8,7 @@ namespace ChatUtils.Contract.Dtos.ChatGroupDto
 {
     public class RecentActiveDto
     {
-        public int RecieverId { get; set; }
+        public long RecieverId { get; set; }
         public bool IsGroup { get; set; }
         public bool IsNull {  get; set; } 
     }
