+++ Dtos/CommunityMessageDto/ChatBoxDto.cs
@@ -10,15 +10,15 @@ namespace ChatUtils.Contract.Dtos.CommunityMessageDto
 {
     public class ChatBoxDto
     {
-        //public int EmployeeId { get; set; }
+        //public long EmployeeId { get; set; }
         //public string? EmployeeName { get; set; }
         public string? LastMessage { get; set; }
         //public bool IsSeen { get; set; }
         public MessageStatus Status { get; set; }
-        public int NewMessages { get; set; }
+        public long NewMessages { get; set; }
         public DateTime? LastMessageDate { get; set; }
 
-        public int RecieverId { get; set; }
+        public long RecieverId { get; set; }
         public string? RecieverName { get; set; }
         public DateTime? LastActive { get; set; }
         public bool IsGroup { get; set; }
