+++ Dtos/PaginationDto/MessagePageDto.cs
@@ -9,8 +9,8 @@ namespace ChatUtils.Contract.Dtos.PaginationDto
     public class MessagePageDto
     {
         public bool IsPagination { get; set; }
-        public int? Index { get; set; }
-        public int? Take { get; set; }
+        public long? Index { get; set; }
+        public long? Take { get; set; }
         public string? Search { get; set; }
     }
 }
