+++ Dtos/PaginationDto/PageDto.cs
@@ -10,11 +10,11 @@ namespace ChatUtils.Contract.Dtos.PaginationDto
     public class PageDto
     {
         public bool  IsPagination { get; set; }
-        public int? Index { get; set; }
-        public int? Take { get; set; }
+        public long? Index { get; set; }
+        public long? Take { get; set; }
         public string? Search {  get; set; }
         public OrdersType? Orders { get; set; }
         public string? OrderBy { get; set; }
-        public List<int>? UserIds { get; set; }
+        public List<long>? UserIds { get; set; }
     }
 }
