+++ ResponseMessage/ResponseMsg.cs
@@ -10,6 +10,6 @@ namespace ChatUtils.Contract.ResponseMessage
     {
         public string Status { get; set; }
         public string Message { get; set; }
-        public int StatusCode { get; set; }
+        public long StatusCode { get; set; }
     }
 }
